+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Versioning follows [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
 ### Added
+- `NexJob.Trigger.RabbitMQ` provider for enqueueing jobs from RabbitMQ messages.
 
 ### Changed